dnsproxy: Add DNS proxying functionality.
There are services that don't use the libc resolver
in their service containers (for example some Go-based
services).

This feature allows the MDNS publisher to act as a
DNS proxy on bridged networks for these services.

Set the 'PROXY_DNS' envvar to 'true' to proxy DNS
instead of publishing MDNS records, and ensure the
networking mode is the bridge network and not 'host'.

Change-type: minor
