+++ config/confd_env_backend/conf.d/env.toml
@@ -8,5 +8,6 @@ keys = [
   "DBUS_SESSION_BUS_ADDRESS",
   "BALENA_SUPERVISOR_ADDRESS",
   "BALENA_SUPERVISOR_API_KEY",
-  "MDNS_API_TOKEN"
+  "MDNS_API_TOKEN",
+  "PROXY_DNS",
 ]
