+++ package.json
@@ -25,11 +25,13 @@
     "bluebird": "^3.5.1",
     "dbus-native": "^0.4.0",
     "lodash": "^4.17.15",
+    "mz": "^2.7.0",
     "request": "^2.88.0",
     "request-promise": "^4.2.4"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.134",
+    "@types/mz": "0.0.32",
     "@types/node": "^10.14.4",
     "@types/request-promise": "^4.1.42",
     "husky": "^1.3.1",
